Move run_workflow to core.py
